Fixed error where qaGbTracks would crash if output from tdbQuery was empty. Included new runCommondNoAbort in the script. Instead of deleting db.sizes file created during process, will instead keep this and utilize it if it already exists which prevents the file from being created and deleted multiple times if qaGbTracks is run on multiple psl tracks. All added at Jonathan's recommendation in CR, refs #18564

+++ python/lib/ucscGb/qa/tables/pslQa.py
@@ -1,4 +1,3 @@
-import subprocess
 import os
 
 from ucscGb.qa.tables.positionalQa import PositionalQa
@@ -15,15 +14,13 @@ class PslQa(PositionalQa):
         # Get baseColorUseSequence setting
         tdbCommand = ["tdbQuery", "select baseColorUseSequence from " + self.db +\
                 " where track='" + self.table +"' or table='"+ self.table +"'"]
-        p = subprocess.Popen(tdbCommand, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        #extract stdout/stderr from Popen command
-        tdbOut, tdbErr = p.communicate()
-        #Turn tdbOut into array so we can extract different pieces
+        tdbOut, tdbErr = qaUtils.runCommand(tdbCommand)
+        # Turn tdbOut into array so we can extract different pieces
         tdbOut = tdbOut.replace("\n","")
         tdbOut = tdbOut.split(" ")
         # tdbOut[1] should be what option is used for baseColorUseSequence
         # We want to fully run pslCheck for those entries with "table" and "extFile"
-        if tdbOut[1] == "table" or tdbOut[1] == "extFile":
+        if len(tdbOut) > 2:
             if tdbOut[1] == "table":
                 # Get sizes of sequences in specified table
                 sizesOut = qaUtils.callHgsql(self.db, "select name, length(seq) from " + tdbOut[2])
@@ -35,30 +32,31 @@ class PslQa(PositionalQa):
             itemSizes = open("%s.sizes" % self.table, 'w')
             itemSizes.write(sizesOut)
             itemSizes.close()
-            # get chrom sizes from chromInfo
-            chromSizesOut = qaUtils.callHgsql(self.db, "select * from chromInfo")
-            chromSizes = open("%s.chrom.sizes" % self.db, 'w')
-            chromSizes.write(chromSizesOut)
-            chromSizes.close()
+            # Check if $db.sizes file exists
+            if not os.path.isfile("%s.chrom.sizes" % self.db):
+                # If it doens't exist, get chrom sizes from chromInfo
+                chromSizesOut = qaUtils.callHgsql(self.db, "select * from chromInfo")
+                chromSizes = open("%s.chrom.sizes" % self.db, 'w')
+                chromSizes.write(chromSizesOut)
+                chromSizes.close()
             # Run more in-depth version of pslCheck
             command = ("pslCheck", "-querySizes=%s.sizes" % self.table ,\
-                    "-targetSizes=%s.chrom.sizes" % self.db, "db=" + self.db, self.table)
+                    "-targetSizes=%s.chrom.sizes" % self.db, "-db=" + self.db, self.table)
             self.reporter.writeCommand(command)
-            p = subprocess.Popen(command, stdout=self.reporter.fh, stderr=self.reporter.fh)
-            pslCheckOut, pslCheckErr = p.communicate()
-            #Clean up intermediate files
+            commandOut, commandErr, commandReturnCode = qaUtils.runCommandNoAbort(command)
+            # Write output to file
+           self.reporter.fh.write(commandErr)
+            # Clean up intermediate item sizes file
             os.remove("%s.sizes" % self.table)
-            os.remove("%s.chrom.sizes" % self.db)
 
         # For everything else, use generic set of steps
         else:
             command = ["pslCheck", "db=" + self.db, self.table]
             self.reporter.writeCommand(command)
-            p = subprocess.Popen(command, stdout=self.reporter.fh, stderr=self.reporter.fh)
-            pslCheckOut, pslCheckErr = p.communicate()
+            commandOut, commandErr, commandReturnCode = qaUtils.runCommandNoAbort(command)
+           self.reporter.fh.write(commandErr)
 
-        p.wait()
-        if p.returncode:
+        if commandReturnCode:
             self.recordError()
         else:
             self.recordPass()
